bot/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
dateutils==0.6.12
Deprecated==1.2.14
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
netaddr==1.2.1
pika==1.3.2
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
yookassa==3.0.1
